#4 Initial commit of driver code.
This patch contains the initial commit for react_loading_screens.py, which is a top-level script for starting a front end server to serve a web application that'll show all of our loading screens. Use the "-start-front-end-server" option for this.

Most of this code is boiler plate.

The "-deploy" option is currently not implemented.

We'll get a NotImplementedError if we try to use it.

There aren't any other top-level CLI options available yet.