#6 Update default page with basic animations
This is a progress patch.

This puts our front end in an intermediate state where some parts of the animation are still not yet completed.

We intend to have a default page with links to all the possible loading animation screens we want to show.

We also intend that there are nice animations on that page as well.

This patch handles some of the CSS complications having to deal with that.

The animations aren't complete; more work is to come.