#8 Initial commit of StingRays component
This patch includes the initial implementation of our first implemented loading screen, i.e. the StingRays component. See react_loading_screens/front_end/src/Components/StingRays/StringRays.*

It's a page with an animation intended to simulate several stingrays moving from left to right across the screen.

There will be some aesthetic improvements coming in the future, e.g. color changes.

This patch also includes quite a bit of refactoring.

Since we intend to have it such that each component is quite modular and can be more-or-less copy-and-pasted into another React project, it would be nice to have the directory structure and the components have a similar modular structure.

This motivates the decisions and changes I'll describe below.

Everything needed to use the StringRays component is in react_loading_screens/front_end/src/Components/StingRays ; just import and use.

We moved what was intended to be the default / index page's content to react_loading_screens/front_end/src/Components/DefaultPageContent/ as we wanted to make this also modular.

Additionally, we abstracted out the stylings relevant to that page from index.scss to DefaultPageContent.scss

Since we intend for everything to be more or less modular in the ways discussed above, we got rid of index.scss as each modular component will be managing its own stylings in its own *.scss file.