#8 Replace StingRays Component with DancingDiamonds Component
After discussion with some third party testers, we've come to the decision that the aesthetics of the StingRays component were lacking. Thus, we've decided to scrap that design and are going for a different one.

This patch changes the aesthetic of the loading screen but came with few changes to the underlying mechanics.

Very few to no changes to the relevant JS code were needed.

A minimal amount of CSS changes were needed (mostly changing the timing functions of animations).

This patch additionally adds some text that is only visible when being passed over by the moving bars, which leads to a nice aesthetic.

There were some font additions and animation additions for the text, but not to the previously existing functionality.