#7 Add link to DancingDiamonds
In an effort to make it so that we can click links on the default / landing page that'll lead us to the different loading screens we have.

This patch adds in the functionality to read the hash in the url that'll take us to the corresponding page.

The only supported corresponding page at this moment is DancingDiamonds.