#7 Use Route instances over hash url information
This patch makes it so that we use BrowserRouter and Route from react-router-dom rather than what we had previously been doing, which is relying on information form the "#" in the url.

The most relevant changes are in App.js

This makes it so that our links from the landing page actually take us to the right loading screens.

Previously, there were issues with that.

It used to only work when we opened in a new tab, which is far from ideal.