#9 Initial Support for PastelQuarters Component
We decided to support a loading screen that consisted of the screens plit into 4 quadrants where each quadrant would switch between some set of pastel colors.

This patch implements that.

The aesthetic results aren't so impressive so far.

As part of #9, we'll continue putting effort into this, but may or may not scrap this screen.