#10 Add Fork & Knife Components
This path is in efforts to complete ticket #10.

We want to make a loading screen inspired by Good Eats.

We decided to create a fork and knife component.

Most of the work was done in CSS and React. See the changes in react_loading_screens/front_end/src/Components/GoodEats