#10 Initial commit of rotating Knife & Fork entrance animation
As part of #10, we are aiming for a food-themed loading screen.

We previously in https://github.com/paul-tqh-nguyen/react_loading_screens/commit/98234cc6b01a8e53e1c01d7b178a5d7f597e421a introduced a Knife and Fork component.

This patch makes it so that the knife and fork components have a rotating entrance in.