#13 Progress patch on initial implementation of "Entering Bars" Loading Screen
This patch is progress on the initial implementation of our "Entering Bars" loading screen.

We want there to be bars entering from the top and left in an alternating fashion.

We want them to exit similarly.

This patch implements an initial working functionality that places all the bars in their correct places that also accounts for z-index.