#21 Make NN depiction's embedding layers to concatenation layer use multiple arrows
The NN depiction is such that each arrow represents one number (and never represents a whole vector or matrix, which isn't the case for all depictions). Thus, it makes sense to have the embedding layer have multiple arrows comming out of them and go into the concatenation layer. Before, there was just one arrow out of each embedding layer, which would indicate that the arrow represented a vector. This is fine for other depictions where we might want to have a more sparse depiection, but it's undesirable in this ccase since every other arrow represents exactly one number.